fix: handle error and return gracefully during failures
The plugin during start, tries to fetch bucket info and also ensure that the bucket exists.
in this process, if the output configs don't have sufficient permissions or privileges to execute any of the gcp apis, it errors with forbidden and fails.

this is problematic in the logging operator context esp., when we allow different teams to setup their log flows with their output sinks and only one of those sinks have an issue. we would expect the rest of the logging ecosystem to work well, while the specific flow fails gracefully without bombing the entire setup